% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{bWoodN1}
\alias{bWoodN1}
\title{Dataset bWoodN1}
\description{
A multivariate dataset (x1, x2, x3, x4, x5, x6, y) of class data.frame and dim 20000 
x 7 to be fitted by a neural network with 5 hidden neurons (41 parameters).
}
\examples{
ht(bWoodN1)
}
\references{
Inspired by page 29 of Wood, S. N. (2011). Fast stable restricted maximum 
likelihood and marginal likelihood estimation of semiparametric generalized linear 
models. Journal of the Royal Statistical Society: Series B (Statistical Methodology), 
73(1), 3-36. 
\url{https://people.bath.ac.uk/man54/SAMBa/ITTs/ITT2/EDF/REMLWood2009.pdf}
}
\keyword{datasets}
