% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clearNN.R
\name{clearNN}
\alias{clearNN}
\alias{detachNN}
\title{Detach the Loaded Packages and the ZZ object}
\usage{
clearNN(donotremove, donotdetach = NULL)

detachNN()
}
\arguments{
\item{donotremove}{a vector of characters representing objects in \code{ls()}.}

\item{donotdetach}{a vector of packages and environments that are not detached.
\code{donotdetach = NULL} protects the packages and environments 
\code{".GlobalEnv"}, \code{"package:kableExtra"}, 
\code{"package:dplyr"}, \code{"package:stringr"}, 
\code{"package:NNbenchmark"}, \code{"package:rmarkdown"}, 
\code{"package:knitr"}, \code{"package:captioner"},
\code{"package:pkgload"}, \code{"package:R6"}, 
\code{"tools:rstudio"}, 
\code{"package:RWsearch"}, \code{"package:pacman"}, 
\code{"package:stats"}, \code{"package:graphics"}, 
\code{"package:grDevices"}, \code{"package:utils"}, 
\code{"package:datasets"}, \code{"package:methods"}, 
\code{"Autoloads"}, \code{"package:base"}.}
}
\description{
\code{clearNN} detachs \code{ZZ} and the packages loaded for the evaluation. 

\code{detachNN} detachs \code{ZZ}.
}
