% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timer_function.R
\name{createTimer}
\alias{createTimer}
\title{Create a timer object}
\usage{
createTimer(verbose = T)
}
\arguments{
\item{verbose}{A parameter to control whether to print messages while using
methods. Default to \code{TRUE}.}
}
\value{
a timer object.
}
\description{
Create a timer object
}
\examples{
## Create, start and stop timing for event 1 (comment is optional)
## Get all records in a data frame
timer1 <- createTimer()      # print is enabled
timer1 <- createTimer(FALSE) # print is disabled
timer1$start("event1")
timer1$stop("event1", RMSE = 1, comment = "event 1 stopped") 
getTimer(timer1)
}
