% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ht.R
\name{ht}
\alias{ht}
\alias{ht9}
\title{Concatenates head() and tail() in vector, list, matrix, data.frame}
\usage{
ht(x, n = 3, m = 4, l = 2)

ht9(x, n = 3, m = 999, l = 2)
}
\arguments{
\item{x}{vectors, matrix, data.frame or list.}

\item{n}{integer. Half-Length of vector or half-numer of lines (for matrix 
and data.frame) of the resulting object.}

\item{m}{integer. Half-number of columns (for matrix and data.frame) of the 
resulting object. \code{m = Inf} prevents any extraction.}

\item{l}{integer. Half-length of the list of the resulting object.}
}
\description{
Concatenates \code{head(n)} and \code{tail(n)} rows and subset with m columns. 
Works also with list.
}
\examples{
### ht
vec <- 1:19
ht(vec)

dfr <- data.frame(A=1:19, B=21:39)
ht(dfr)

mat <- matrix(1:200, ncol = 20)
ht9(mat)


lst <- list(A = 11:13, B = 21:59, C = 31:69, 
            D = 41:43, E = 51:88, G = 61:99, 
            H = matrix(1:300, ncol = 15, dimnames = list(1:20, 1:15)))
ht(lst)
ht(lst, n = 2, m = 3, l = 2)
ht9(lst, l = 2)

}
