% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{mDette}
\alias{mDette}
\title{Dataset mDette}
\description{
A multivariate dataset (x1, x2, x3, y) of class matrix and dim 500 x 4 to be 
fitted by a neural network with 5 hidden neurons (26 parameters).
}
\examples{
ht(mDette)
pairs(mDette)
}
\references{
Dette, H., & Pepelyshev, A. (2010). Generalized Latin hypercube design for computer 
experiments. Technometrics, 52(4). 

See also \url{https://www.sfu.ca/~ssurjano/detpep10curv.html}
}
\keyword{datasets}
