% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{mFriedman}
\alias{mFriedman}
\title{Dataset mFriedman}
\description{
A multivariate dataset (x1, x2, x3, x4, x5, y) of class matrix and dim 500 x 6 to be  
fitted by a neural network with 4 or 5 hidden neurons.
}
\examples{
ht(mFriedman)
pairs(mFriedman)
}
\references{
Friedman, J. H., Grosse, E., & Stuetzle, W. (1983). Multidimensional additive spline 
approximation. SIAM Journal on Scientific and Statistical Computing, 4(2), 291-301. 

See also \url{http://www.sfu.ca/~ssurjano/fried.html}
}
\keyword{datasets}
