% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{mIshigami}
\alias{mIshigami}
\title{Dataset mIshigami}
\description{
A multivariate dataset (x1, x2, x3, y) of class matrix and dim 500 x 4 to be fitted 
by a neural network with 10 hidden neurons (51 parameters).
}
\examples{
ht(mIshigami)
pairs(mIshigami)
}
\references{
Ishigami, T., & Homma, T. (1990, December). An importance quantification technique in 
uncertainty analysis for computer models. In Uncertainty Modeling and Analysis, 1990. 
Proceedings., First International Symposium on (pp. 398-403). IEEE. 

See also \url{https://www.sfu.ca/~ssurjano/ishigami.html}
}
\keyword{datasets}
