% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{mRef153}
\alias{mRef153}
\title{Dataset mRef153}
\description{
A multivariate dataset (x1, x2, x3, x4, x5, y) of class matrix and dim 153 x 6 to be 
fitted by a neural network with 3 hidden neurons (22 parameters).
This dataset was used to teach neural networks at ESPCI from 2003 to 2013 and is 
available in the software Neuro One.
}
\examples{
ht(mRef153)
pairs(mRef153)
}
\references{
Neuro One \url{https://www.inmodelia.com/software.html}
}
\keyword{datasets}
