% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{NNbigdatasets}
\alias{NNbigdatasets}
\title{Big Datasets in One list (2020)}
\description{
\code{NNbigdatasets} is a list with the big datasets presented in this package and the 
recommended number of hidden neurons for each neural network model. 

\itemize{
  \item{bWoodN1:    5 neurons.}
}

Each item of the list is itself a list with 5 components:
\itemize{
\item{ds: character. The name of the dataset.}
\item{neur: integer. The recommanded number of hidden neurons in the NN model and in 
\code{fmlaNN}.}
\item{nparNN: integer. The number of parameters in \code{fmlaNN}.}
\item{fmlaNN: the formula of the corresponding neural network, with tanh() as the
activation function in the hidden layer.}
\item{Z: matrix or data.frame. The dataset itself.}
}
Using \code{attach()} and \code{detach()} gives a direct access to these items.
}
\examples{
ht(NNbigdatasets)

NNdataSummary(NNbigdatasets)

}
\keyword{datasets}
