% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NNdataSummary.R
\name{NNdataSummary}
\alias{NNdataSummary}
\title{A Summary (in data.frame format) of NNdatasets}
\usage{
NNdataSummary(NNdatasets)
}
\arguments{
\item{NNdatasets}{the NNdatasets list.}
}
\value{
A data.frame with 12 rows and 5 columns: (dataset) name, n_rows, n_inputs, 
n_neurons, n_parameters.
}
\description{
\code{NNdataSummary} summarizes the information of the 12 datasets
listed in \code{NNdatasets}.
}
\examples{
NNdataSummary(NNdatasets)

}
