% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNN.R
\name{plotNN}
\alias{plotNN}
\alias{lipoNN}
\title{Create a Plot, Add Lines or Points Depending The Context}
\usage{
plotNN(xory, y0, uni, TF = TRUE, ...)

lipoNN(xory, y_pred, uni, TF = TRUE, ...)
}
\arguments{
\item{xory}{vector of numeric. The original x values for an univariate dataset 
or the original y values for a multivariate dataset.}

\item{y0}{vector of numeric. The original y values.}

\item{uni}{logicial. TRUE for an univariate dataset. FALSE for a multivariate dataset.
\code{...} fails or if you call the function from another function.}

\item{TF}{logical. TRUE executes the instruction. FALSE ignores the instruction.
Equivalent to \code{if (TRUE/FALSE) plot()}.}

\item{...}{parameters passed to \code{plot()}, \code{lines()} or \code{points()}.}

\item{y_pred}{vector of numeric. The values returned by the \code{predict()} function.}
}
\value{
NULL in the console. An initial plot or some added lines/points.
}
\description{
\code{plotNN} uses the parameter \code{uni} to launch the \code{plot()} function either 
for a univariate dataset (x, y_pred) or for a multivariate dataset (y, y_pred).

\code{lipoNN} uses the parameter \code{uni} to launch either \code{lines()} for an
univariate dataset or \code{points()} for a multivariate dataset.

See the examples in \code{\link{prepareZZ}}.
}
