% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestart-timediff.R
\name{timestart}
\alias{timestart}
\alias{timediff}
\title{Collect the difftime between 2 intervals}
\usage{
timestart()

timediff()
}
\value{
A single numeric value that represents a duration in seconds.
}
\description{

}
\examples{
timestart()
Sys.sleep(2)
timediff()

}
