% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{uDmod1}
\alias{uDmod1}
\title{Dataset uDmod1}
\description{
An univariate dataset (x, y) of class data.frame and dim 51 x 2 to be fitted by a 
neural network with 6 hidden neurons (19 parameters).
The parameters are highly correlated and singular Jacobian matrices often appear. 
A difficult dataset.
}
\examples{
ht(uDmod1)
plot(uDmod1)
}
\keyword{datasets}
