% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{uDmod2}
\alias{uDmod2}
\title{Dataset uDmod2}
\description{
An univariate dataset (x, y) of class data.frame and dim 51 x 2 to be fitted by a 
neural network with 5 hidden neurons (16 parameters).
}
\examples{
ht(uDmod2)
plot(uDmod2)
}
\keyword{datasets}
