% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{uDreyfus1}
\alias{uDreyfus1}
\title{Dataset uDreyfus1}
\description{
An univariate dataset (x, y) of class data.frame and dim 51 x 2 to be fitted by a 
neural network with 3 hidden neurons (10 parameters).
This dataset was used to teach neural networks at ESPCI from 1991 to 2013. It usually
appeared in the very first slides. 
This is a combination of 3 pure tanh() functions without noise. The Jacobian matrix 
is singular at the target parameter values and many algorithms could fail.
}
\examples{
ht(uDreyfus1)
plot(uDreyfus1)
}
\references{
Dreyfus, G., ESPCI \url{https://www.neurones.espci.fr}
}
\keyword{datasets}
