% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{uGauss1}
\alias{uGauss1}
\title{Dataset uGauss1}
\description{
An univariate dataset (x, y) of class data.frame and dim 250 x 2 to be fitted by a 
neural network with 5 hidden neurons (16 parameters).
}
\examples{
ht(uGauss1)
plot(uGauss1)
}
\references{
Rust, B., NIST (1996) \url{https://www.inmodelia.com/gsoc2020-redirection.html}
with redirection to the (slow) NIST page.
}
\keyword{datasets}
