% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{uGauss3}
\alias{uGauss3}
\title{Dataset uGauss3}
\description{
An univariate dataset (x, y) of class data.frame and dim 250 x 2 to be fitted by a 
neural network with 4 hidden neurons (13 parameters).
}
\examples{
ht(uGauss3)
plot(uGauss3)
}
\references{
Rust, B., NIST (1996) \url{https://www.inmodelia.com/gsoc2020-redirection.html}
with redirection to the (slow) NIST page.
}
\keyword{datasets}
