% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{uNeuroOne}
\alias{uNeuroOne}
\title{Dataset uNeuroOne}
\description{
An univariate dataset (x, y) of class data.frame and dim 51 x 2 to be fitted by a 
neural network with 2 hidden neurons (7 parameters).
This dataset was used to teach neural networks at ESPCI from 1991 to 2013 and is 
available in the software Neuro One.
}
\examples{
ht(uNeuroOne)
plot(uNeuroOne)
}
\references{
Dreyfus, G., ESPCI \url{https://www.neurones.espci.fr}

Neuro One \url{https://www.inmodelia.com/software.html}
}
\keyword{datasets}
