% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add2csv.R
\name{add2csv}
\alias{add2csv}
\title{Create or Append a data.frame To a csv File}
\usage{
add2csv(x, file = "results.csv")
}
\arguments{
\item{x}{a data.frame or a matrix.}

\item{file}{character. The filename.}
}
\description{
Create or append a data.frame to a csv file. Column names are added at creation and 
ignored at the append steps.
}
\examples{
x <- data.frame(a = 1:4, b = 5:8)  
add2csv(x) 
add2csv(x) 
add2csv(x)
read.csv("results.csv")

}
