% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clearNN.R
\name{clearNN}
\alias{clearNN}
\title{Clean the .GlobalEnv and Detach the Loaded Packages}
\usage{
clearNN(donotremove, donotdetach = NULL)
}
\arguments{
\item{donotremove}{a vector of characters representing objects in \code{ls()}.}

\item{donotdetach}{a vector of packages and environments that are not detached.
\code{donotdetach = NULL} protects the packages and environments 
\code{".GlobalEnv"}, \code{"package:NNbenchmark"}, 
\code{"package:rmarkdown"}, \code{"package:knitr"}, \code{"package:captioner"},
\code{"package:pkgload"}, \code{"package:R6"}, \code{"tools:rstudio"}, 
\code{"package:RWsearch"}, \code{"package:pacman"}, 
\code{"package:stats"}, \code{"package:graphics"}, 
\code{"package:grDevices"}, \code{"package:utils"}, 
\code{"package:datasets"}, \code{"package:methods"}, 
\code{"Autoloads"}, \code{"package:base"}.}
}
\description{
Clean the first environment \code{ls(.GlobalEnv)} from objects not listed in 
\code{donotremove} and detach the loaded packages and their dependencies not listed 
in \code{donotdetach}.

\code{clearNN()} replaces

\code{detach(ZZ)} 

\code{detach("package:nameNNpackage", unload = TRUE)} 

\code{remove(list = ls()[!(ls() \%in\% donotremove)])}
}
