% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funRMSE.R
\name{funRMSE}
\alias{funRMSE}
\alias{funMAE}
\title{Calculate the RMSE and MAE, Round to 4 digits}
\usage{
funRMSE(y_pred, y0, dgts = 4)

funMAE(y_pred, y0, dgts = 4)
}
\arguments{
\item{y_pred}{vector of numeric. The predicted values.}

\item{y0}{vector of numeric. The original values.}

\item{dgts}{integer. The rounding.}
}
\description{
Calculate the Root Means Square Error (RMSE) and the Mean Absoluter Error (MAE)
and round the result, by default to 4 digits. Apply \code{na.rm = TRUE}
}
\examples{
y0 <- 1:19
y_pred <- y0 + rnorm(length(y0), sd = 0.3)
funRMSE(y_pred, y0)
funMAE( y_pred, y0)

}
