% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_intro.R
\docType{package}
\name{NNbenchmark-package}
\alias{NNbenchmark-package}
\alias{NNbenchmark}
\title{Package NNbenchmark}
\description{
Datasets and functions to benchmark neural network functions and training algorithms 
(convergence, speed, ease of use) in 69 R packages.
}
\examples{
ds <- grep("^[m,u]", ls("package:NNbenchmark"), value = TRUE); ds
t(sapply(ds, function(x) dim(get(x))))
plot(uGauss2)
pairs(mIshigami)
ht(NNdatasets, n = 2, l = 6)
}
