% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{NNdatasets}
\alias{NNdatasets}
\title{All Datasets in One List}
\description{
\code{NNdatasets} is a list with the 12 datasets presented in this package and the 
recommanded number of hidden neurons for each neural network model. 

\itemize{
  \item{mDette:    5 neurons.}
  \item{mFriedman: 4 or 5 neurons.}
  \item{mIshigami: 10 neurons.}
  \item{mRef153:   3 neurons.}
  \item{uDmod1:    5 or 6 neurons.}
  \item{uDmod2:    5 neurons.}
  \item{uDreyfus1: 3 neurons.}
  \item{uDreyfus2: 3 neurons.}
  \item{uGauss1:   5 neurons.}
  \item{uGauss2:   4 neurons.}
  \item{uGauss3:   4 neurons.}
  \item{uNeuroOne: 3 neurons.}
}

Each item of the list is itself a list with 3 components:
\itemize{
\item{ds: character. The name of the dataset.}
\item{neur: integer. The number of hidden neurons in the NN model.}
\item{Z: matrix or data.frame. The dataset itself.}
}
Using \code{attach()} and \code{detach()} gives a direct access to these items.
}
\examples{
ht(NNdatasets, n = 2, l = 6)

for (i in names(NNdatasets)) {
    DS <- NNdatasets[[i]]
    attach(DS)
    print(ds)
    print(neur)
    print(ht(Z, n = 2))
    cat("\\n")
    ## Type here the code to test the 69 neural network packages.
    detach(DS)
}

}
\keyword{datasets}
