% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timer_class.R
\docType{class}
\name{timeR}
\alias{timeR}
\title{A R6 Class to represent a timer.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
timeR
}
\description{
timer is a R6 Class that represent a timer. This is a modified version of the 
\code{timeR} package for an internal use. Full credit is to Yifu Yan, the author 
of the \code{timeR} package.
}
\details{
\code{getTimer} returns a data frame with all records saved by the timer 
object. Columns in the data.frame are: event, start, end, duration, RMSE, 
MAE, stars, params, comment.
}
\section{Fields}{

\describe{
\item{\code{time}}{A POSIXct/POSIXlt value of your latest timing.}

\item{\code{event}}{A string of your latest timing.}

\item{\code{eventTable}}{A data frame that stores all timings.}

\item{\code{verbose}}{A printing setting that controls whether to print messages.}
}}

\section{Public Methods}{

\describe{
  \item{\code{initialize(time,event,verbose,eventTable)}
  }{Initialize a timer object. You can also use \code{createTimer()}
  function to initialize a timer object.}
  \item{\code{start(eventName)}
  }{Start timing for a event, \code{eventName} should be a string}
  \item{\code{stop(eventName)}
  }{Stop timing for a event.}
  \item{\code{getTimer()}
  }{Get/Print a data.frame with all records.}
  \item{\code{removeEvent(eventName)}
  }{Remove an given row in the eventTable.}
  \item{\code{toggleVerbose()}
  }{Toggle between \code{TRUE} and \code{FALSE} for \code{verbose}}
  \item{\code{getStartTime()}
  }{Get start time for a selected event.}
  \item{\code{getStopTime()}
  }{Get stop time for a selected event.}
  \item{\code{getDuration()}
  }{Get duration for a selected event.}
  \item{\code{getRMSE()}
  }{Get the RMSE for a selected event.}
  \item{\code{getMAE()}
  }{Get the MAE for a selected event.}
  \item{\code{getStars()}
  }{Get stars for a selected event.}
  \item{\code{getParams()}
  }{Get params for a selected event.}
  \item{\code{getComment()}
  }{Get comment for a selected event.}
  \item{\code{getEventf()}
  }{Get entire row for a selected event.}
  \item{\code{print()}
  }{Custom print method for timer class. However, you don't need to use this
  function to generate custom printing.
  Custom printing is triggered by default.}
  }
}

\section{Private Methods}{

\describe{
\item{\code{slprint(msg, flag = self$verbose)}
}{A function that controls whether to print extra message.}
}
}

\examples{
timer <- createTimer()
timer$start("event1")
# put some codes in between, for instance
Sys.sleep(1)
timer$stop("event1", RMSE = 1, MAE = 1.3, stars = "*", 
           params = "maxiter=100, lr=0.01", comment = "OK for 1",  
           printmsg = TRUE)

timer$start("event2")
# put some codes in between, for instance
Sys.sleep(2)
timer$stop("event2", RMSE = 2, MAE = 2.6, stars = "**",  
           params = "maxiter=1000, lr=0.001", comment = "OK for 2",  
           printmsg = FALSE)

table1 <- getTimer(timer)
timer$toggleVerbose() # set verbose to FALSE as default is TRUE

table1 # print all records in a tibble(data frame)

# get attributes for selected events
timer$getStartTime("event1")
timer$getStopTime("event1")
timer$getDuration("event1")
timer$getComment("event1")
timer$getEvent("event1")
}
\keyword{datasets}
