% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestart-timediff.R
\name{timestart}
\alias{timestart}
\alias{timediff}
\title{Collect the difftime between two events}
\usage{
timestart()

timediff()
}
\value{
A single numeric value that represents a duration in seconds.
}
\description{
\code{timestart} starts the timer and saved the value in an object named 
\code{time0} stored in \code{.GlobalEnv}.

\code{timediff} stops the timer, remove the \code{time0} objet from \code{.GlobalEnv}
and prints the duration in seconds between the two events.

\code{timestart} and \code{timediff} are fully independant from  the R6 class
\code{timeR} and the objects \code{createTimer} or \code{getTimer}. They use
\code{\link{proc.time}} instead.
}
\examples{
timestart()
Sys.sleep(2)
timediff()

}
